update
add results to user model
