+++ models/test.js
@@ -1,5 +1,4 @@
 const { Schema, model } = require("mongoose");
-const Joi = require("joi");
 const handleSaveErrors = require("../middlewares/handleSaveErrors");
 
 const testSchema = new Schema(
@@ -30,21 +29,8 @@ const testSchema = new Schema(
 
 testSchema.post("save", handleSaveErrors);
 
-const getSchema = Joi.object({
-  question: Joi.string().required(),
-  questionId: Joi.number().required(),
-  answers: Joi.array().required(),
-  rightAnswer: Joi.string().required(),
-  type: Joi.string().required(),
-});
-
-const schemas = {
-  getSchema,
-};
-
 const Test = model("test", testSchema);
 
 module.exports = {
   Test,
-  schemas,
 };
