+++ models/users.js
@@ -18,6 +18,10 @@ const userSchema = new Schema(
       type: String,
       default: "",
     },
+    results: {
+      type: Number,
+      default: 0,
+    },
     // avatarURL: {
     //   type: String,
     //   required: true,
