+++ schemas/index.js
@@ -1,3 +1,4 @@
 const registerSchema = require("./registerSchema");
+const getSchema = require("./registerSchema");
 
-module.exports = { registerSchema };
+module.exports = { registerSchema, getSchema };
